Make Sass no longer required
